# docs/openapi/generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@livetale.com
    name: LiveTale Team
  description: |
    LiveTale �꾨줈�앺듃�� API 紐낆꽭��
    �명꽣�숉떚釉� �ㅽ넗由ы뀛留곴낵 罹먮┃�� 梨꾪똿�� �꾪븳 醫낇빀 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LiveTale API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.livetale.com/v1
- description: Staging server
  url: https://staging-api.livetale.com/v1
- description: Development server
  url: http://localhost:3000/v1
security:
- BearerAuth: []
- ApiKeyAuth: []
paths:
...
  /mypage/profile:
    get:
      description: �ъ슜�먯쓽 �곸꽭 �꾨줈�꾧낵 怨꾩젙 �뺣낫瑜� 議고쉶�⑸땲��
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_mypage_profile_get_200_response"
          description: �꾨줈�� �뺣낫 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �꾨줈��/怨꾩젙 �뺣낫
      tags:
      - MyPage
      x-accepts:
      - application/json
      x-tags:
      - tag: MyPage
  /mypage/stories:
    get:
      description: �ъ슜�먭� 吏꾪뻾 以묒씠嫄곕굹 �꾨즺�� �ㅽ넗由� 紐⑸줉�� 議고쉶�⑸땲��
      parameters:
      - description: �ㅽ넗由� �곹깭 �꾪꽣
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - active
          - paused
          - completed
          - abandoned
          type: string
        style: form
      - description: �섏씠吏� 踰덊샇
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: �섏씠吏��� ��ぉ ��
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 50
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_mypage_stories_get_200_response"
          description: �ㅽ넗由� 紐⑸줉 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �닿� 吏꾪뻾/�꾧껐�� �ㅽ넗由� 由ъ뒪��
      tags:
      - MyPage
      x-accepts:
      - application/json
      x-tags:
      - tag: MyPage
  /mypage/usage:
    get:
      description: �ъ슜�먯쓽 �꾩옱 �ъ슜�됯낵 �붽툑�� �뺣낫瑜� 議고쉶�⑸땲��
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_mypage_usage_get_200_response"
          description: �ъ슜�� �뺣낫 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �ъ슜�� / �붽툑�� �곹깭
      tags:
      - MyPage
      x-accepts:
      - application/json
      x-tags:
      - tag: MyPage
  /mypage/relationships:
    get:
      description: �ъ슜�먯� 罹먮┃�곕뱾 媛꾩쓽 愿�怨� �붿빟�� 議고쉶�⑸땲��
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_mypage_relationships_get_200_response"
          description: 愿�怨� �붿빟 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: 愿�怨� �붿빟
      tags:
      - MyPage
      x-accepts:
      - application/json
      x-tags:
      - tag: MyPage
...
components:
  schemas:
    User:
      example:
        subscription_tier: free
        is_active: true
        avatar_url: https://openapi-generator.tech
        updated_at: 2000-01-23T04:56:07.000+00:00
        nickname: nickname
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          description: �ъ슜�� 怨좎쑀 ID
          format: uuid
          type: string
        email:
          description: �대찓�� 二쇱냼
          format: email
          type: string
        username:
          description: �ъ슜�먮챸
          maxLength: 20
          minLength: 3
          type: string
        nickname:
          description: �됰꽕��
          maxLength: 30
          type: string
        avatar_url:
          description: �꾨줈�� �대�吏� URL
          format: uri
          type: string
        created_at:
          description: 怨꾩젙 �앹꽦��
          format: date-time
          type: string
        updated_at:
          description: 留덉�留� �섏젙��
          format: date-time
          type: string
        is_active:
          description: 怨꾩젙 �쒖꽦�� �곹깭
          type: boolean
        subscription_tier:
          description: 援щ룆 �깃툒
          enum:
          - free
          - premium
          - pro
          type: string
      type: object
...
    Character:
      example:
        world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        appearance: appearance
        personality: personality
        avatar_url: https://openapi-generator.tech
        name: name
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        world_id:
          format: uuid
          type: string
        name:
          maxLength: 50
          minLength: 1
          type: string
        description:
          maxLength: 500
          type: string
        personality:
          maxLength: 1000
          type: string
        appearance:
          maxLength: 1000
          type: string
        avatar_url:
          format: uri
          type: string
        created_at:
          format: date-time
          type: string
      type: object
    Story:
      example:
        world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        updated_at: 2000-01-23T04:56:07.000+00:00
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        current_scene_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: active
      properties:
        id:
          format: uuid
          type: string
        world_id:
          format: uuid
          type: string
        title:
          type: string
        description:
          type: string
        current_scene_id:
          format: uuid
          type: string
        status:
          enum:
          - active
          - paused
          - completed
          - abandoned
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object
...
    Achievement:
      example:
        name: name
        description: description
        progress: 8
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        category: category
        unlocked_at: 2000-01-23T04:56:07.000+00:00
        is_completed: true
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        progress:
          maximum: 100
          minimum: 0
          type: integer
        is_completed:
          type: boolean
        unlocked_at:
          format: date-time
          type: string
      type: object
...
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      type: object
...
    _worlds_get_200_response_pagination:
      example:
        total: 1
        limit: 6
        page: 0
        total_pages: 5
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      type: object
...
    _mypage_profile_get_200_response_stats:
      example:
        characters_created: 6
        days_active: 5
        total_chat_messages: 1
        stories_completed: 0
      properties:
        stories_completed:
          description: �꾨즺�� �ㅽ넗由� ��
          type: integer
        characters_created:
          description: �앹꽦�� 罹먮┃�� ��
          type: integer
        total_chat_messages:
          description: 珥� 梨꾪똿 硫붿떆吏� ��
          type: integer
        days_active:
          description: �쒖꽦 �ъ슜 �쇱닔
          type: integer
      type: object
    _mypage_profile_get_200_response_achievements_summary:
      example:
        recent_achievements:
        - name: name
          description: description
          progress: 8
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          category: category
          unlocked_at: 2000-01-23T04:56:07.000+00:00
          is_completed: true
        - name: name
          description: description
          progress: 8
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          category: category
          unlocked_at: 2000-01-23T04:56:07.000+00:00
          is_completed: true
        completed_achievements: 2
        total_achievements: 5
      properties:
        total_achievements:
          type: integer
        completed_achievements:
          type: integer
        recent_achievements:
          items:
            $ref: "#/components/schemas/Achievement"
          type: array
      type: object
    _mypage_profile_get_200_response:
      example:
        stats:
          characters_created: 6
          days_active: 5
          total_chat_messages: 1
          stories_completed: 0
        user:
          subscription_tier: free
          is_active: true
          avatar_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          nickname: nickname
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
        achievements_summary:
          recent_achievements:
          - name: name
            description: description
            progress: 8
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            category: category
            unlocked_at: 2000-01-23T04:56:07.000+00:00
            is_completed: true
          - name: name
            description: description
            progress: 8
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            category: category
            unlocked_at: 2000-01-23T04:56:07.000+00:00
            is_completed: true
          completed_achievements: 2
          total_achievements: 5
      properties:
        user:
          $ref: "#/components/schemas/User"
        stats:
          $ref: "#/components/schemas/_mypage_profile_get_200_response_stats"
        achievements_summary:
          $ref: "#/components/schemas/_mypage_profile_get_200_response_achievements_summary"
      type: object
    _mypage_stories_get_200_response_stories_inner:
      allOf:
      - $ref: "#/components/schemas/Story"
      - properties:
          progress_percentage:
            maximum: 100
            minimum: 0
            type: integer
          last_played:
            format: date-time
            type: string
          world_name:
            type: string
        type: object
      example:
        world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        world_name: world_name
        updated_at: 2000-01-23T04:56:07.000+00:00
        progress_percentage: 8
        last_played: 2000-01-23T04:56:07.000+00:00
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        current_scene_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: active
    _mypage_stories_get_200_response:
      example:
        stories:
        - world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          world_name: world_name
          updated_at: 2000-01-23T04:56:07.000+00:00
          progress_percentage: 8
          last_played: 2000-01-23T04:56:07.000+00:00
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          current_scene_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: active
        - world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          world_name: world_name
          updated_at: 2000-01-23T04:56:07.000+00:00
          progress_percentage: 8
          last_played: 2000-01-23T04:56:07.000+00:00
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          current_scene_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: active
        pagination:
          total: 1
          limit: 6
          page: 0
          total_pages: 5
      properties:
        stories:
          items:
            $ref: "#/components/schemas/_mypage_stories_get_200_response_stories_inner"
          type: array
        pagination:
          $ref: "#/components/schemas/_worlds_get_200_response_pagination"
      type: object
    _mypage_usage_get_200_response_subscription:
      example:
        expires_at: 2000-01-23T04:56:07.000+00:00
        tier: free
        status: active
      properties:
        tier:
          enum:
          - free
          - premium
          - pro
          type: string
        status:
          enum:
          - active
          - expired
          - cancelled
          type: string
        expires_at:
          format: date-time
          type: string
      type: object
    _mypage_usage_get_200_response_usage:
      example:
        stories_created: 0
        chat_messages_sent: 1
        characters_created: 6
        images_generated: 5
      properties:
        stories_created:
          description: �앹꽦�� �ㅽ넗由� ��
          type: integer
        characters_created:
          description: �앹꽦�� 罹먮┃�� ��
          type: integer
        chat_messages_sent:
          description: �꾩넚�� 梨꾪똿 硫붿떆吏� ��
          type: integer
        images_generated:
          description: �앹꽦�� �대�吏� ��
          type: integer
      type: object
    _mypage_usage_get_200_response_limits:
      example:
        max_stories: 5
        max_characters: 2
        monthly_images: 9
        monthly_chat_messages: 7
      properties:
        max_stories:
          description: 理쒕� �ㅽ넗由� ��
          type: integer
        max_characters:
          description: 理쒕� 罹먮┃�� ��
          type: integer
        monthly_chat_messages:
          description: �붽컙 梨꾪똿 硫붿떆吏� �쒕룄
          type: integer
        monthly_images:
          description: �붽컙 �대�吏� �앹꽦 �쒕룄
          type: integer
      type: object
    _mypage_usage_get_200_response:
      example:
        usage:
          stories_created: 0
          chat_messages_sent: 1
          characters_created: 6
          images_generated: 5
        subscription:
          expires_at: 2000-01-23T04:56:07.000+00:00
          tier: free
          status: active
        limits:
          max_stories: 5
          max_characters: 2
          monthly_images: 9
          monthly_chat_messages: 7
      properties:
        subscription:
          $ref: "#/components/schemas/_mypage_usage_get_200_response_subscription"
        usage:
          $ref: "#/components/schemas/_mypage_usage_get_200_response_usage"
        limits:
          $ref: "#/components/schemas/_mypage_usage_get_200_response_limits"
      type: object
    _mypage_relationships_get_200_response_character_relationships_inner:
      example:
        character:
          world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          appearance: appearance
          personality: personality
          avatar_url: https://openapi-generator.tech
          name: name
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        relationship_level: 8
        favorite_topics:
        - favorite_topics
        - favorite_topics
        last_interaction: 2000-01-23T04:56:07.000+00:00
        interaction_count: 6
      properties:
        character:
          $ref: "#/components/schemas/Character"
        relationship_level:
          description: 愿�怨� �섏� (0-100)
          maximum: 100
          minimum: 0
          type: integer
        interaction_count:
          description: �곹샇�묒슜 �잛닔
          type: integer
        last_interaction:
          format: date-time
          type: string
        favorite_topics:
          items:
            type: string
          type: array
      type: object
    _mypage_relationships_get_200_response_most_interacted:
      example:
        character:
          world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          appearance: appearance
          personality: personality
          avatar_url: https://openapi-generator.tech
          name: name
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        interaction_count: 5
      properties:
        character:
          $ref: "#/components/schemas/Character"
        interaction_count:
          type: integer
      type: object
    _mypage_relationships_get_200_response:
      example:
        character_relationships:
        - character:
            world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            appearance: appearance
            personality: personality
            avatar_url: https://openapi-generator.tech
            name: name
            description: description
            created_at: 2000-01-23T04:56:07.000+00:00
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          relationship_level: 8
          favorite_topics:
          - favorite_topics
          - favorite_topics
          last_interaction: 2000-01-23T04:56:07.000+00:00
          interaction_count: 6
        - character:
            world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            appearance: appearance
            personality: personality
            avatar_url: https://openapi-generator.tech
            name: name
            description: description
            created_at: 2000-01-23T04:56:07.000+00:00
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          relationship_level: 8
          favorite_topics:
          - favorite_topics
          - favorite_topics
          last_interaction: 2000-01-23T04:56:07.000+00:00
          interaction_count: 6
        most_interacted:
          character:
            world_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            appearance: appearance
            personality: personality
            avatar_url: https://openapi-generator.tech
            name: name
            description: description
            created_at: 2000-01-23T04:56:07.000+00:00
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          interaction_count: 5
        total_characters: 1
      properties:
        character_relationships:
          items:
            $ref: "#/components/schemas/_mypage_relationships_get_200_response_character_relationships_inner"
          type: array
        total_characters:
          description: 珥� �곹샇�묒슜�� 罹먮┃�� ��
          type: integer
        most_interacted:
          $ref: "#/components/schemas/_mypage_relationships_get_200_response_most_interacted"
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey