# src/main/java/com/LiveToon/api/MypageApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.LiveToon.api;

import com.LiveToon.model.Error;
import com.LiveToon.model.MypageProfileGet200Response;
import com.LiveToon.model.MypageRelationshipsGet200Response;
import com.LiveToon.model.MypageStoriesGet200Response;
import com.LiveToon.model.MypageUsageGet200Response;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-09-29T19:10:41.347886+09:00[Asia/Seoul]", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "MyPage", description = "the MyPage API")
public interface MypageApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_MYPAGE_PROFILE_GET = "/mypage/profile";
    /**
     * GET /mypage/profile : �꾨줈��/怨꾩젙 �뺣낫
     * �ъ슜�먯쓽 �곸꽭 �꾨줈�꾧낵 怨꾩젙 �뺣낫瑜� 議고쉶�⑸땲��
     *
     * @return �꾨줈�� �뺣낫 議고쉶 �깃났 (status code 200)
     *         or �몄쬆�섏� �딆� �ъ슜�� (status code 401)
     */
    @Operation(
        operationId = "mypageProfileGet",
        summary = "�꾨줈��/怨꾩젙 �뺣낫",
        description = "�ъ슜�먯쓽 �곸꽭 �꾨줈�꾧낵 怨꾩젙 �뺣낫瑜� 議고쉶�⑸땲��",
        tags = { "MyPage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "�꾨줈�� �뺣낫 議고쉶 �깃났", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MypageProfileGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "�몄쬆�섏� �딆� �ъ슜��", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MypageApi.PATH_MYPAGE_PROFILE_GET,
        produces = { "application/json" }
    )
    
    default ResponseEntity<MypageProfileGet200Response> mypageProfileGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"stats\" : { \"characters_created\" : 6, \"days_active\" : 5, \"total_chat_messages\" : 1, \"stories_completed\" : 0 }, \"user\" : { \"subscription_tier\" : \"free\", \"is_active\" : true, \"avatar_url\" : \"https://openapi-generator.tech\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"nickname\" : \"nickname\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"email\" : \"email\", \"username\" : \"username\" }, \"achievements_summary\" : { \"recent_achievements\" : [ { \"name\" : \"name\", \"description\" : \"description\", \"progress\" : 8, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"category\" : \"category\", \"unlocked_at\" : \"2000-01-23T04:56:07.000+00:00\", \"is_completed\" : true }, { \"name\" : \"name\", \"description\" : \"description\", \"progress\" : 8, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"category\" : \"category\", \"unlocked_at\" : \"2000-01-23T04:56:07.000+00:00\", \"is_completed\" : true } ], \"completed_achievements\" : 2, \"total_achievements\" : 5 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"{}\", \"error\" : \"error\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_MYPAGE_RELATIONSHIPS_GET = "/mypage/relationships";
    /**
     * GET /mypage/relationships : 愿�怨� �붿빟
     * �ъ슜�먯� 罹먮┃�곕뱾 媛꾩쓽 愿�怨� �붿빟�� 議고쉶�⑸땲��
     *
     * @return 愿�怨� �붿빟 議고쉶 �깃났 (status code 200)
     *         or �몄쬆�섏� �딆� �ъ슜�� (status code 401)
     */
    @Operation(
        operationId = "mypageRelationshipsGet",
        summary = "愿�怨� �붿빟",
        description = "�ъ슜�먯� 罹먮┃�곕뱾 媛꾩쓽 愿�怨� �붿빟�� 議고쉶�⑸땲��",
        tags = { "MyPage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "愿�怨� �붿빟 議고쉶 �깃났", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MypageRelationshipsGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "�몄쬆�섏� �딆� �ъ슜��", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MypageApi.PATH_MYPAGE_RELATIONSHIPS_GET,
        produces = { "application/json" }
    )
    
    default ResponseEntity<MypageRelationshipsGet200Response> mypageRelationshipsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"character_relationships\" : [ { \"character\" : { \"world_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"appearance\" : \"appearance\", \"personality\" : \"personality\", \"avatar_url\" : \"https://openapi-generator.tech\", \"name\" : \"name\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"relationship_level\" : 8, \"favorite_topics\" : [ \"favorite_topics\", \"favorite_topics\" ], \"last_interaction\" : \"2000-01-23T04:56:07.000+00:00\", \"interaction_count\" : 6 }, { \"character\" : { \"world_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"appearance\" : \"appearance\", \"personality\" : \"personality\", \"avatar_url\" : \"https://openapi-generator.tech\", \"name\" : \"name\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"relationship_level\" : 8, \"favorite_topics\" : [ \"favorite_topics\", \"favorite_topics\" ], \"last_interaction\" : \"2000-01-23T04:56:07.000+00:00\", \"interaction_count\" : 6 } ], \"most_interacted\" : { \"character\" : { \"world_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"appearance\" : \"appearance\", \"personality\" : \"personality\", \"avatar_url\" : \"https://openapi-generator.tech\", \"name\" : \"name\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"interaction_count\" : 5 }, \"total_characters\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"{}\", \"error\" : \"error\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_MYPAGE_STORIES_GET = "/mypage/stories";
    /**
     * GET /mypage/stories : �닿� 吏꾪뻾/�꾧껐�� �ㅽ넗由� 由ъ뒪��
     * �ъ슜�먭� 吏꾪뻾 以묒씠嫄곕굹 �꾨즺�� �ㅽ넗由� 紐⑸줉�� 議고쉶�⑸땲��
     *
     * @param status �ㅽ넗由� �곹깭 �꾪꽣 (optional)
     * @param page �섏씠吏� 踰덊샇 (optional, default to 1)
     * @param limit �섏씠吏��� ��ぉ �� (optional, default to 20)
     * @return �ㅽ넗由� 紐⑸줉 議고쉶 �깃났 (status code 200)
     *         or �몄쬆�섏� �딆� �ъ슜�� (status code 401)
     */
    @Operation(
        operationId = "mypageStoriesGet",
        summary = "�닿� 吏꾪뻾/�꾧껐�� �ㅽ넗由� 由ъ뒪��",
        description = "�ъ슜�먭� 吏꾪뻾 以묒씠嫄곕굹 �꾨즺�� �ㅽ넗由� 紐⑸줉�� 議고쉶�⑸땲��",
        tags = { "MyPage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "�ㅽ넗由� 紐⑸줉 議고쉶 �깃났", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MypageStoriesGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "�몄쬆�섏� �딆� �ъ슜��", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MypageApi.PATH_MYPAGE_STORIES_GET,
        produces = { "application/json" }
    )
    
    default ResponseEntity<MypageStoriesGet200Response> mypageStoriesGet(
        @Parameter(name = "status", description = "�ㅽ넗由� �곹깭 �꾪꽣", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) @Nullable String status,
        @Min(1) @Parameter(name = "page", description = "�섏씠吏� 踰덊샇", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(50) @Parameter(name = "limit", description = "�섏씠吏��� ��ぉ ��", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"stories\" : [ { \"world_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"world_name\" : \"world_name\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"progress_percentage\" : 8, \"last_played\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"title\" : \"title\", \"current_scene_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"active\" }, { \"world_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"world_name\" : \"world_name\", \"updated_at\" : \"2000-01-23T04:56:07.000+00:00\", \"progress_percentage\" : 8, \"last_played\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"created_at\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"title\" : \"title\", \"current_scene_id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"active\" } ], \"pagination\" : { \"total\" : 1, \"limit\" : 6, \"page\" : 0, \"total_pages\" : 5 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"{}\", \"error\" : \"error\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_MYPAGE_USAGE_GET = "/mypage/usage";
    /**
     * GET /mypage/usage : �ъ슜�� / �붽툑�� �곹깭
     * �ъ슜�먯쓽 �꾩옱 �ъ슜�됯낵 �붽툑�� �뺣낫瑜� 議고쉶�⑸땲��
     *
     * @return �ъ슜�� �뺣낫 議고쉶 �깃났 (status code 200)
     *         or �몄쬆�섏� �딆� �ъ슜�� (status code 401)
     */
    @Operation(
        operationId = "mypageUsageGet",
        summary = "�ъ슜�� / �붽툑�� �곹깭",
        description = "�ъ슜�먯쓽 �꾩옱 �ъ슜�됯낵 �붽툑�� �뺣낫瑜� 議고쉶�⑸땲��",
        tags = { "MyPage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "�ъ슜�� �뺣낫 議고쉶 �깃났", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MypageUsageGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "�몄쬆�섏� �딆� �ъ슜��", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MypageApi.PATH_MYPAGE_USAGE_GET,
        produces = { "application/json" }
    )
    
    default ResponseEntity<MypageUsageGet200Response> mypageUsageGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"usage\" : { \"stories_created\" : 0, \"chat_messages_sent\" : 1, \"characters_created\" : 6, \"images_generated\" : 5 }, \"subscription\" : { \"expires_at\" : \"2000-01-23T04:56:07.000+00:00\", \"tier\" : \"free\", \"status\" : \"active\" }, \"limits\" : { \"max_stories\" : 5, \"max_characters\" : 2, \"monthly_images\" : 9, \"monthly_chat_messages\" : 7 } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"{}\", \"error\" : \"error\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}